+++ examples/resources/iosxe_bgp_ipv4_unicast_vrf_neighbor/resource.tf
@@ -14,6 +14,7 @@ resource "iosxe_bgp_ipv4_unicast_vrf_neighbor" "example" {
   timers_minimum_neighbor_hold              = 30
   version                                   = 4
   fall_over_default_route_map               = "RMAP"
+  fall_over_bfd                             = true
   fall_over_bfd_single_hop                  = true
   fall_over_bfd_check_control_plane_failure = true
   fall_over_bfd_strict_mode                 = true