+++ internal/provider/model_iosxe_bgp_ipv4_unicast_vrf_neighbor.go
@@ -53,6 +53,7 @@ type BGPIPv4UnicastVRFNeighbor struct {
 	TimersMinimumNeighborHold           types.Int64                          `tfsdk:"timers_minimum_neighbor_hold"`
 	Version                             types.Int64                          `tfsdk:"version"`
 	FallOverDefaultRouteMap             types.String                         `tfsdk:"fall_over_default_route_map"`
+	FallOverBfd                         types.Bool                           `tfsdk:"fall_over_bfd"`
 	FallOverBfdMultiHop                 types.Bool                           `tfsdk:"fall_over_bfd_multi_hop"`
 	FallOverBfdSingleHop                types.Bool                           `tfsdk:"fall_over_bfd_single_hop"`
 	FallOverBfdCheckControlPlaneFailure types.Bool                           `tfsdk:"fall_over_bfd_check_control_plane_failure"`
@@ -92,6 +93,7 @@ type BGPIPv4UnicastVRFNeighborData struct {
 	TimersMinimumNeighborHold           types.Int64                          `tfsdk:"timers_minimum_neighbor_hold"`
 	Version                             types.Int64                          `tfsdk:"version"`
 	FallOverDefaultRouteMap             types.String                         `tfsdk:"fall_over_default_route_map"`
+	FallOverBfd                         types.Bool                           `tfsdk:"fall_over_bfd"`
 	FallOverBfdMultiHop                 types.Bool                           `tfsdk:"fall_over_bfd_multi_hop"`
 	FallOverBfdSingleHop                types.Bool                           `tfsdk:"fall_over_bfd_single_hop"`
 	FallOverBfdCheckControlPlaneFailure types.Bool                           `tfsdk:"fall_over_bfd_check_control_plane_failure"`
@@ -180,6 +182,11 @@ func (data BGPIPv4UnicastVRFNeighbor) toBody(ctx context.Context) string {
 	}
 	if !data.FallOverDefaultRouteMap.IsNull() && !data.FallOverDefaultRouteMap.IsUnknown() {
 		body, _ = sjson.Set(body, helpers.LastElement(data.getPath())+"."+"fall-over.default.route-map", data.FallOverDefaultRouteMap.ValueString())
+	}
+	if !data.FallOverBfd.IsNull() && !data.FallOverBfd.IsUnknown() {
+		if data.FallOverBfd.ValueBool() {
+			body, _ = sjson.Set(body, helpers.LastElement(data.getPath())+"."+"fall-over.bfd", map[string]string{})
+		}
 	}
 	if !data.FallOverBfdMultiHop.IsNull() && !data.FallOverBfdMultiHop.IsUnknown() {
 		if data.FallOverBfdMultiHop.ValueBool() {
@@ -348,6 +355,15 @@ func (data *BGPIPv4UnicastVRFNeighbor) updateFromBody(ctx context.Context, res g
 		data.FallOverDefaultRouteMap = types.StringValue(value.String())
 	} else {
 		data.FallOverDefaultRouteMap = types.StringNull()
+	}
+	if value := res.Get(prefix + "fall-over.bfd"); !data.FallOverBfd.IsNull() {
+		if value.Exists() {
+			data.FallOverBfd = types.BoolValue(true)
+		} else {
+			data.FallOverBfd = types.BoolValue(false)
+		}
+	} else {
+		data.FallOverBfd = types.BoolNull()
 	}
 	if value := res.Get(prefix + "fall-over.bfd.multi-hop"); !data.FallOverBfdMultiHop.IsNull() {
 		if value.Exists() {
@@ -564,6 +580,11 @@ func (data *BGPIPv4UnicastVRFNeighborData) fromBody(ctx context.Context, res gjs
 	if value := res.Get(prefix + "fall-over.default.route-map"); value.Exists() {
 		data.FallOverDefaultRouteMap = types.StringValue(value.String())
 	}
+	if value := res.Get(prefix + "fall-over.bfd"); value.Exists() {
+		data.FallOverBfd = types.BoolValue(true)
+	} else {
+		data.FallOverBfd = types.BoolValue(false)
+	}
 	if value := res.Get(prefix + "fall-over.bfd.multi-hop"); value.Exists() {
 		data.FallOverBfdMultiHop = types.BoolValue(true)
 	} else {
@@ -691,6 +712,9 @@ func (data *BGPIPv4UnicastVRFNeighbor) getDeletedItems(ctx context.Context, stat
 	if !state.FallOverDefaultRouteMap.IsNull() && data.FallOverDefaultRouteMap.IsNull() {
 		deletedItems = append(deletedItems, fmt.Sprintf("%v/fall-over/default/route-map", state.getPath()))
 	}
+	if !state.FallOverBfd.IsNull() && data.FallOverBfd.IsNull() {
+		deletedItems = append(deletedItems, fmt.Sprintf("%v/fall-over/bfd", state.getPath()))
+	}
 	if !state.FallOverBfdMultiHop.IsNull() && data.FallOverBfdMultiHop.IsNull() {
 		deletedItems = append(deletedItems, fmt.Sprintf("%v/fall-over/bfd/multi-hop", state.getPath()))
 	}
@@ -781,6 +805,9 @@ func (data *BGPIPv4UnicastVRFNeighbor) getEmptyLeafsDelete(ctx context.Context) 
 	if !data.LogNeighborChangesDisable.IsNull() && !data.LogNeighborChangesDisable.ValueBool() {
 		emptyLeafsDelete = append(emptyLeafsDelete, fmt.Sprintf("%v/log-neighbor-changes/disable", data.getPath()))
 	}
+	if !data.FallOverBfd.IsNull() && !data.FallOverBfd.ValueBool() {
+		emptyLeafsDelete = append(emptyLeafsDelete, fmt.Sprintf("%v/fall-over/bfd", data.getPath()))
+	}
 	if !data.FallOverBfdMultiHop.IsNull() && !data.FallOverBfdMultiHop.ValueBool() {
 		emptyLeafsDelete = append(emptyLeafsDelete, fmt.Sprintf("%v/fall-over/bfd/multi-hop", data.getPath()))
 	}
@@ -852,6 +879,9 @@ func (data *BGPIPv4UnicastVRFNeighbor) getDeletePaths(ctx context.Context) []str
 	}
 	if !data.FallOverDefaultRouteMap.IsNull() {
 		deletePaths = append(deletePaths, fmt.Sprintf("%v/fall-over/default/route-map", data.getPath()))
+	}
+	if !data.FallOverBfd.IsNull() {
+		deletePaths = append(deletePaths, fmt.Sprintf("%v/fall-over/bfd", data.getPath()))
 	}
 	if !data.FallOverBfdMultiHop.IsNull() {
 		deletePaths = append(deletePaths, fmt.Sprintf("%v/fall-over/bfd/multi-hop", data.getPath()))