+++ internal/provider/resource_iosxe_bgp_neighbor_test.go
@@ -36,6 +36,7 @@ func TestAccIosxeBGPNeighbor(t *testing.T) {
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_bgp_neighbor.test", "disable_connected_check", "false"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_bgp_neighbor.test", "fall_over_default_enable", "false"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_bgp_neighbor.test", "fall_over_default_route_map", "RMAP"))
+	checks = append(checks, resource.TestCheckResourceAttr("iosxe_bgp_neighbor.test", "fall_over_bfd", "true"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_bgp_neighbor.test", "fall_over_bfd_single_hop", "true"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_bgp_neighbor.test", "fall_over_bfd_check_control_plane_failure", "true"))
 	checks = append(checks, resource.TestCheckResourceAttr("iosxe_bgp_neighbor.test", "fall_over_bfd_strict_mode", "true"))
@@ -102,6 +103,7 @@ func testAccIosxeBGPNeighborConfig_all() string {
 	config += `	disable_connected_check = false` + "\n"
 	config += `	fall_over_default_enable = false` + "\n"
 	config += `	fall_over_default_route_map = "RMAP"` + "\n"
+	config += `	fall_over_bfd = true` + "\n"
 	config += `	fall_over_bfd_single_hop = true` + "\n"
 	config += `	fall_over_bfd_check_control_plane_failure = true` + "\n"
 	config += `	fall_over_bfd_strict_mode = true` + "\n"